Update facemesh.py
Added two functions:

face_swap_vid() which takes in an image as parameter. If we pass in the webcam image as a parameter it will swap face with Will Smiths's face frame by frame.

The other function does the same thing but on video. It is a void function, no need to pass in an image. I added two images, the will smith image and a target image. We can add more images later.